requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-CORS
python-dotenv==0.21.0
pytz==2023.3
ultralytics==8.0.101
torch>=1.13.1
opencv-python==4.7.0.72
scikit-image==0.20.0
scikit-learn==1.2.2
pandas==2.0.3
Flask-Migrate==4.0.4
psutil==5.9.5
Pillow==9.5.0
PyMuPDF==1.24.0
matplotlib==3.7.1
gunicorn==20.1.0
